Started an Introduction
In this stage, I decided a detailed question related to Question 1 and started Introduction part. I briefly summarized the dataset I would use and the exact variables I would use for this project. I haven't found the reference about my background. And in the next stage, I decided to come up with the methods to upload and clean the data.